Store prior failures in '.failed' scratch directory
When a combination of gathered requirements fails save the
contents of that failure in a scratch directory with the
requirements that composed that failure (using the md5 of the
requirements contents as the filename) and use this in later
checks to see if we should bother continuing down a prior
checked requirement list/set. When we have found a prior set
that has already failed just abort early (rather than later).